pyproject.toml
[project]
name = "edgee"
version = "1.0.0"
description = "Lightweight Python SDK for Edgee AI Gateway"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.10"
dependencies = ["pydantic>=2.0.0"]
authors = [
    { name = "Edgee", email = "support@edgee.cloud" }
]
keywords = ["ai", "llm", "gateway", "openai", "anthropic"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/edgee-cloud/python-sdk"
Repository = "https://github.com/edgee-cloud/python-sdk"

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.8.0", "python-dotenv>=1.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "python-dotenv>=1.0.0",
]
